refactor: Split ContactFields in multiple files
Changes : FieldValue was renamed in FieldWrapper. renderFieldValue has been converted in func comp and was renamed in FieldByType
