+++ src/components/ContactCard/ContactCard.jsx
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types'
 import { translate } from 'cozy-ui/transpiled/react/I18n'
 import ContactIdentity from './ContactIdentity'
 import ContactGroupsList from '../ContactGroups/ContactGroupList'
-import ContactFields from './ContactFields'
+import ContactFields from './ContactFields/ContactFields'
 import { fullContactPropTypes } from '../ContactPropTypes'
 import { normalizeFields, getConnectedAccounts } from '../../helpers/contacts'
 import ContactAccounts from './ContactAccounts'
